# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for get cryptocurrencies",
        "title": "My-Portfolio",
        "contact": {
            "name": "My-Portfolio support"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "paths": {
...
        "/api/v1/stakings": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "список стейкингов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staking"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_app_controllers.StakingGetResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/stakings/detail/{crypt_symbol}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "список стейкингов для конкретной криптовалюты",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staking"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cryptocurrency symbol",
                        "name": "crypt_symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_app_controllers.StakingDetailGetResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/stakings/portfolio": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "список стейкингов для криптовалют в портфеле",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "staking"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_app_controllers.StakingGetResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_app_controllers.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "fl_my-portfolio_internal_clients.CryptocurrencyDataDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quote": {
                    "type": "object",
                    "properties": {
                        "BTC": {
                            "type": "object",
                            "properties": {
                                "percent_change_24h": {
                                    "description": "Volume24h        float64 `json:\"volume_24h\"`\nVolumeChange24h  float64 `json:\"volume_change_24h\"`\nPercentChange1h  float64 `json:\"percent_change_1h\"`",
                                    "type": "number"
                                },
                                "percent_change_30d": {
                                    "description": "PercentChange7d  float64 `json:\"percent_change_7d\"`",
                                    "type": "number"
                                },
                                "percent_change_90d": {
                                    "description": "PercentChange60d float64 `json:\"percent_change_60d\"`",
                                    "type": "number"
                                },
                                "price": {
                                    "type": "number"
                                }
                            }
                        },
                        "RUB": {
                            "type": "object",
                            "properties": {
                                "percent_change_24h": {
                                    "description": "Volume24h        float64 `json:\"volume_24h\"`\nVolumeChange24h  float64 `json:\"volume_change_24h\"`\nPercentChange1h  float64 `json:\"percent_change_1h\"`",
                                    "type": "number"
                                },
                                "percent_change_30d": {
                                    "description": "PercentChange7d  float64 `json:\"percent_change_7d\"`",
                                    "type": "number"
                                },
                                "percent_change_90d": {
                                    "description": "PercentChange60d float64 `json:\"percent_change_60d\"`",
                                    "type": "number"
                                },
                                "price": {
                                    "type": "number"
                                }
                            }
                        },
                        "USD": {
                            "type": "object",
                            "properties": {
                                "percent_change_24h": {
                                    "description": "Volume24h        float64 `json:\"volume_24h\"`\nVolumeChange24h  float64 `json:\"volume_change_24h\"`\nPercentChange1h  float64 `json:\"percent_change_1h\"`",
                                    "type": "number"
                                },
                                "percent_change_30d": {
                                    "description": "PercentChange7d  float64 `json:\"percent_change_7d\"`",
                                    "type": "number"
                                },
                                "percent_change_90d": {
                                    "description": "PercentChange60d float64 `json:\"percent_change_60d\"`",
                                    "type": "number"
                                },
                                "price": {
                                    "type": "number"
                                }
                            }
                        }
                    }
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "fl_my-portfolio_internal_clients.StakingData": {
            "type": "object",
            "properties": {
                "apy": {
                    "type": "number"
                },
                "asset": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "quota_minimum": {
                    "type": "number"
                },
                "reward_asset": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_controllers.ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "error_string": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_controllers.StakingDetailGetResponse": {
            "type": "object",
            "properties": {
                "binance_stakings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fl_my-portfolio_internal_clients.StakingData"
                    }
                },
                "bybit_stakings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fl_my-portfolio_internal_clients.StakingData"
                    }
                },
                "okx_stakings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fl_my-portfolio_internal_clients.StakingData"
                    }
                }
            }
        },
        "internal_app_controllers.StakingGetResponse": {
            "type": "object",
            "properties": {
                "binance_staking": {
                    "$ref": "#/definitions/fl_my-portfolio_internal_clients.StakingData"
                },
                "bybit_staking": {
                    "$ref": "#/definitions/fl_my-portfolio_internal_clients.StakingData"
                },
                "cryptocurrency_data": {
                    "$ref": "#/definitions/fl_my-portfolio_internal_clients.CryptocurrencyDataDetail"
                },
                "okx_staking": {
                    "$ref": "#/definitions/fl_my-portfolio_internal_clients.StakingData"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}